requirements.txt
discord.py>=2.3,<3.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
aiosqlite>=0.19,<1.0
humanize>=4.9,<5.0
